Modified how DALEC variables are fed into LAI_KNORR
The DALEC variables ET and PAW, along with the LAI_KNORR
variables that need to be stored in memory and passed onto the
next iteration, are now set at the end of the DALEC time loop.
So, essentially the current iteration of the loop takes in these
variables from the previous iteration.
Also made minor change in LAI_KNORR for lambda_W.