Temporary change in dLAI/dt form
There is some issue arising from a difference in analytical
(incremental) and numerical forms of LAI rate of change. The
incremental form seems to be working fine but this does not
match with the original calculation of dLAI/dt, so for now
we set dLAI/dt to be the incremental difference between the new
LAI and the previous LAI (need to reconcile this later).