Fixed error in previous commit
Two things: (1) the negative sign was in the wrong place in the
foliar to litter carbon flux calculation. (2) the calculation of
the minimum of requested C for dlambdadt (LAI change) and available
labile carbon did not consider the fact that one is a flux
(g C m-2 day-1) and one is pool size (g C m-2), so the comparison
ignored the fact that the pool size needs to be per day - to fix
this I divided the labile C pool by deltat.