Added correction to Knorr LAI if labile C limits
If labile C is too small to support new LAI growth (in C flux units)
then we down-scale the dlambda/dt and knorr LAI. THis ensures that
labile C does not go negative. We also added a flag in the FLUXES
output to indicate when labile C limits LAI growth.