Coded LAI fire losses and dynamic LMCA calcs - not correct yet!
I have added in a calculation of the fire losses from the LAI (now
that Cfoliar and LAI are decoupled this needs to be done in addition
to the Cfoliar fire losses). Also, I've included a variable for the
dynamic LMCA which is included in the foliar carbon inputs/outputs
and downscaling of LAI calculations - this doesn't appear to be
working totally correctly yet. The NEE from the fluxes does not
balance the NEE from the delta(POOLS).