Fixed bug in carbon balance
The bug noted in the last commit has been corrected. The "leaf production"
flux should be zero in this model as direct allocation from NPP to foliar C
is no longer considered. Additionally, the ODE for foliar C should use the
labile release flux, not the leaf production flux, as its input flux. The
rate of change in the C pools and net fluxes now balance (to within expected
machine precision 10e-6).