Modified Knorr LAI par range and setup test MDF script
Set the parameters min and max for the Knorr model to be more
realistic. Also modified the field selected for LAI in the computation
of the cost function to the Knorr LAI. The test script added here
also does everything needed to optimize the model against the default
site, considering only LAI observations.