Changed plant-available water timestep input for LAI
Because plant-available water (PAW) is a model pool state, computed
to represent the beginning and end of each month, I have set the input
of PAW into the Knorr model to represent the average state across
the month. Therefore, the average of the previous state and next
state. This makes it consistent with the ET and MET data used in
Knorr LAI.