Adjusted time stepping units in LAI_KNORR.c - no substantial effect on outputs
I modified the LAI_KNORR.c module so that there is no assumption of time-step
size in there. Every rate change variable should now be in units
per day, not per model timestep. That is with the exception of dlambdadt,
which is the change in LAI over the model timestep (m2/m2), so this needs to be
divided by deltat to get the LAI rate of change (m2/m2/day)