Removed beta_factor on leaf dark/day respiration in LIU_AN_ET_REFACTOR.c
Removed the beta_factor and temperature scaling of Rd, the leaf
dark/day respiration rate in LIU_AN_ET_REFACTOR.c. As it was, any
water stress imposed on Rd via the beta_factor will have decreased
the rate of leaf respiration which is the opposite of what is observed
(doi: 10.1093/jxb/erv492; doi: 10.1007/s11099-013-0026-1). Temperature
scaling will be introduced as a separate function next.