Edit ALLOC_AND_AUTO_RESP_FLUXES module
The formulation of the growth respiration and actual growth
functions had some inconsistencies, and the true meaning of the
parameter rauto_gr may have been incorrect. I have modified the code
and re-defined the parameter for growth respiration (rauto_gr) to be
consistent with the "growth yield" term (Y_g) in Cannell and Thornley
(2000, doi: 10.1006/anbo.1999.0996).