Modified Knorr LAI phenology for use in 1100: only KNORR_ALLOCATION.c
Updated the Knorr LAI phenology module used in model ID 1100 to take
transpiration (transp) as an input, instead of evapotranspiration (ET).
The use of ET was there from development with ID 1025, which does not
model transpiration and evaporation separately. This formulation now
used in 1100 (module: KNORR_ALLOCATION.c) is consistent with the
original formulation by Knorr et al. (2010)